Thread-local support for multiple modules. (#3538)
The main module uses -ftls-model=local-exec flag.
This is the most optimal. The compiler/linker hard-codes the offsets from FS
since it knows that the main module is the first one to be loaded.

The secondary modules are compiles with -ftls-model=initial-exec flag.
This informs the compiler that the modules will be loaded along with the main module,
rather than being dynamically loaded via dlopen. In this case, the compiler
will generate R_X86_64_TPOFF relocations for thread-local variables.
Each thread-local variable has a corresponding offset variable. In order to
access a thread-local variable, the value of the corresponding offset variable is
first read and then added to FS. Since thread-locals have negative offsets, this
implies that the offset variables have negative values.

The loader places the thread local sections (tdata + tbss) of the modules one after
the other, starting with the main module. The linker initializes the offset variables
to offsets within the module's tls section. The loader converts these offsets to values
to values relative to the start of the aggregated tls section. Care is taken to account
for possible different alignments of the different tls sections.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>